+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Add the following flags in [`ProgressionCallback`](https://poutyne.org/callbacks.html#poutyne.ProgressionCallback):
   `show_every_n_train_steps`, `show_every_n_valid_steps`, `show_every_n_test_steps`. They allow to show only certain
   steps instead of all steps.
+* Fix bug where all warnings were silenced.
+* Add `strict` flag when loading checkpoints. In Model, a NamedTuple is returned as in PyTorch's `load_state_dict`. In
+  Experiment, a warning is raised when there are missing or unexpected keys in the checkpoint.
 
 # v1.3.1